put return dataclasses above their callers
