+++ colour/difference/cam02_ucs.py
@@ -25,10 +25,11 @@ import numpy as np
 if typing.TYPE_CHECKING:
     from colour.hints import NDArrayFloat, Literal
 
-from colour.difference.typing import DeltaEJabData
+from dataclasses import dataclass, field
+
 from colour.hints import Domain100  # noqa: TC001
 from colour.models.cam02_ucs import COEFFICIENTS_UCS_LUO2006, Coefficients_UCS_Luo2006
-from colour.utilities import as_float, tsplit
+from colour.utilities import MixinDataclassArithmetic, as_float, tsplit
 
 __author__ = "Colour Developers"
 __copyright__ = "Copyright 2013 Colour Developers"
@@ -43,6 +44,46 @@ __all__ = [
     "delta_E_CAM02SCD",
     "delta_E_CAM02UCS",
 ]
+
+
+@dataclass
+class DeltaEData(MixinDataclassArithmetic):
+    """
+    Colour difference data containing the colour difference :math:`\\Delta E`.
+
+    This class is the base container for colour difference computation results
+    returned when ``additional_data=True`` in :mod:`colour.difference` functions.
+    """
+
+    dE: NDArrayFloat | None = field(default_factory=lambda: None)
+
+
+@dataclass
+class DeltaEJabData(DeltaEData):
+    """
+    Colour difference data expressed in a :math:`J'a'b'` colourspace.
+
+    This data structure is returned by the following functions when
+    ``additional_data=True``:
+
+    - :func:`colour.difference.delta_E_Luo2006`
+    - :func:`colour.difference.delta_E_CAM02LCD`
+    - :func:`colour.difference.delta_E_CAM02SCD`
+    - :func:`colour.difference.delta_E_CAM02UCS`
+
+    Notes
+    -----
+    The interpretation of the components is as follows:
+
+    - ``dJ`` is weighted by the lightness coefficient :math:`K_L`
+      defined by the selected *Luo et al. (2006)* uniform colourspace.
+    - ``da`` and ``db`` are raw differences in the corresponding
+      *CAM02-LCD*, *CAM02-SCD* or *CAM02-UCS* colourspace.
+    """
+
+    dJ: NDArrayFloat | None = field(default_factory=lambda: None)
+    da: NDArrayFloat | None = field(default_factory=lambda: None)
+    db: NDArrayFloat | None = field(default_factory=lambda: None)
 
 
 @typing.overload
